Remove some more unused code and dependencies
