+++ package.json
@@ -59,7 +59,6 @@
     "@types/react": "^18.2.61",
     "@types/react-dom": "^18.2.19",
     "canvas-confetti": "^1.9.1",
-    "client-oauth2": "^4.3.3",
     "dotenv": "^16.4.5",
     "eslint-config-next": "^14.1.3",
     "jsdom": "^24.0.0",
@@ -69,7 +68,6 @@
     "next": "^14.1.3",
     "next-auth": "^4.24.6",
     "nodemailer": "^6.9.10",
-    "patch-package": "^8.0.0",
     "pg": "^8.11.3",
     "react": "^18.2.0",
     "react-aria": "^3.32.1",
