+++ tsconfig.json
@@ -96,8 +96,5 @@
   "exclude": [
     "node_modules/",
     "sentry.*.config.ts",
-    // TODO NEXT.JS MIGRATION: Remove old Express files:
-    "src/controllers/**/*.js",
-    "src/views/**/*.js"
   ]
 }
